+++ qir/qsharp-target-tests/conftest.py
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 #
-from target1 import *
+from TargetTestScripts import *
 
 
 __all__ = target1_tests + target3_tests + target4_tests
